requirements.txt
# ─── InfiniteClaw Dependencies ───
# Core
streamlit>=1.50.0
typer[all]>=0.9.0
rich>=13.0.0
litellm>=1.0.0
openai>=1.0.0
python-dotenv>=1.0.0

# SSH & Remote
paramiko>=3.0.0
requests>=2.30.0

# Data & Utils
pandas>=2.0.0
Pillow>=10.0.0
PyYAML>=6.0.0
psutil>=5.9.0
duckduckgo-search>=6.0.0

# Security
cryptography>=41.0.0

# Channels
python-telegram-bot>=21.0

# Desktop
pywebview>=5.0.0

# Server
fastapi>=0.100.0
uvicorn>=0.23.0
